Fix tests to work with new package layout
Unit tests does not compile and run properly

Tests are needed to work properly in order to know that the software
works to some extent

Let's fix the unit tests to make sure they compile and run properly

JUnit is used to run the tests
