+++ src/test/java/jank/task/TaskListTest.java
@@ -1,4 +1,4 @@
-package jank;
+package jank.task;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -7,9 +7,6 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.junit.jupiter.api.Test;
-
-import jank.task.TaskList;
-import jank.task.TodoTask;
 
 
 public class TaskListTest {
